Added Warhammer of Even
The Warhammer of Even is a weapon that can be devastating or
completely useless depending on the opponent's armor. If their
armor has an even value, the warhammer does 85 points of damage,
however if the armor is an odd number, the hitPoints will be divided
by half of the opponent's armor value rendering it nearly useless.